packages/requirements-locked.txt
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.11.0
Authlib==1.6.4
babel==2.17.0
bandit==1.8.6
bcrypt==5.0.0
black==25.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
coverage==7.10.7
cryptography==46.0.2
cycler==0.12.1
docutils==0.21.2
dparse==0.6.4
ecdsa==0.19.1
fastapi==0.118.0
filelock==3.19.1
flake8==7.3.0
fonttools==4.60.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
imagesize==1.4.1
iniconfig==2.1.0
isort==6.1.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.0.1
matplotlib==3.10.6
mccabe==0.7.0
mdurl==0.1.2
mypy==1.18.2
mypy_extensions==1.1.0
nltk==3.9.2
numpy==2.3.3
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
psutil==7.1.0
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rich==14.1.0
roman-numerals-py==3.1.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
safety==3.6.2
safety-schemas==0.0.16
scipy==1.16.2
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
starlette==0.48.0
stevedore==5.5.0
tenacity==9.1.2
tomlkit==0.13.3
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
